+++ apps/ledger-live-desktop/src/renderer/families/ethereum/StakeFlowModal/EthStakingModal.tsx
@@ -1,13 +1,11 @@
 import { Flex } from "@ledgerhq/react-ui";
 import { Account } from "@ledgerhq/types-live";
-import React, { useMemo } from "react";
+import React from "react";
 import { Trans } from "react-i18next";
 import TrackPage from "~/renderer/analytics/TrackPage";
 import Modal, { ModalBody } from "~/renderer/components/Modal";
-import { EthStakingModalBody as EthStakingModalBodyV1 } from "./v1/EthStakingModalBody";
-import { EthStakingModalBody as EthStakingModalBodyV2 } from "./v2/EthStakingModalBody";
-import { ListProviders, ProvidersV1 } from "./types";
-import { getV2ListProviders } from "./utils/getV2ListProviders";
+import { ListProviders } from "./types";
+import { EthStakingModalBody } from "./EthStakingModalBody";
 
 type Props = {
   name: string;
@@ -26,8 +24,6 @@ export function EthStakingModal({
   source,
   listProviders = [],
 }: Props) {
-  const v2Providers = useMemo(() => getV2ListProviders(listProviders), [listProviders]);
-
   return (
     <Modal
       name={name}
@@ -40,24 +36,14 @@ export function EthStakingModal({
           render={() => (
             <Flex justifyContent={"center"}>
               <TrackPage category="ETH Stake Modal" name="Main Modal" />
-              {v2Providers.length ? (
-                <EthStakingModalBodyV2
-                  onClose={onClose}
-                  account={account}
-                  source={source}
-                  listProviders={v2Providers}
-                />
-              ) : (
-                <EthStakingModalBodyV1
-                  onClose={onClose}
-                  account={account}
-                  checkbox={checkbox}
-                  singleProviderRedirectMode={singleProviderRedirectMode}
-                  source={source}
-                  // safe to assume that all providers here are v1.
-                  listProviders={listProviders as ProvidersV1}
-                />
-              )}
+              <EthStakingModalBody
+                onClose={onClose}
+                account={account}
+                checkbox={checkbox}
+                singleProviderRedirectMode={singleProviderRedirectMode}
+                source={source}
+                listProviders={listProviders}
+              />
             </Flex>
           )}
         />
