+++ apps/ledger-live-desktop/src/types/featureFlags.ts
@@ -1,17 +1,10 @@
 export type EthStakingProviders = {
-  listProvider: Array<
-    | {
-        name: string;
-        liveAppId: string;
-        supportLink: string;
-      }
-    | {
-        id: string;
-        name: string;
-        liveAppId: string;
-        supportLink?: string;
-        icon?: string;
-        queryParams?: Record<string, string>;
-      }
-  >;
+  listProvider: Array<{
+    id: string;
+    name: string;
+    liveAppId: string;
+    supportLink?: string;
+    icon?: string;
+    queryParams?: Record<string, string>;
+  }>;
 };
